Move memory params to memtx options table. Issue #1927 #1268
Move memory options to memtx options table. Allow k/m/g suffixes for
memory params. Encode memory limits in bytes.

+++ src/box/tuple.h
@@ -748,7 +748,7 @@ tuple_to_buf(const struct tuple *tuple, char *buf, size_t size);
 
 /** Initialize tuple library */
 void
-tuple_init(float alloc_arena_max_size, uint32_t slab_alloc_minimal,
+tuple_init(uint64_t alloc_arena_max_size, uint32_t slab_alloc_minimal,
 	   uint32_t slab_alloc_maximal, float alloc_factor);
 
 /** Cleanup tuple library */
